Refactor code to fix Norminette
