+++ philo/actions.c
@@ -6,73 +6,23 @@
 /*   By: ribana-b <ribana-b@student.42malaga.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/06 00:00:14 by ribana-b          #+#    #+# Malaga      */
-/*   Updated: 2024/07/07 00:16:15 by ribana-b         ###   ########.com      */
+/*   Updated: 2024/07/07 00:39:54 by ribana-b         ###   ########.com      */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philosophers.h"
 
-void	take_single_fork(t_philo *philo, int side)
-{
-	pthread_mutex_lock(&philo->mutex);
-	philo->fork_taken[side] = true;
-	pthread_mutex_unlock(&philo->mutex);
-}
-
-void	leave_single_fork(t_philo *philo, int side)
-{
-	pthread_mutex_lock(&philo->mutex);
-	philo->fork_taken[side] = false;
-	pthread_mutex_unlock(&philo->mutex);
-}
-
 void	take_forks(t_philo *philo)
 {
 	if (philo->id % 2 == 0)
 	{
-		pthread_mutex_lock(&philo->info->table.fork[philo->fork[R]]);
-		take_single_fork(philo, R);
-		pthread_mutex_lock(&philo->info->mutex);
-		if (quick_check(philo))
-		{
-			pthread_mutex_unlock(&philo->info->mutex);
+		if (safe_take_fork_even(philo))
 			return ;
-		}
-		pthread_mutex_unlock(&philo->info->mutex);
-		print_message(philo, TAKING_FORK, false);
-		pthread_mutex_lock(&philo->info->table.fork[philo->fork[L]]);
-		take_single_fork(philo, L);
-		pthread_mutex_lock(&philo->info->mutex);
-		if (quick_check(philo))
-		{
-			pthread_mutex_unlock(&philo->info->mutex);
-			return ;
-		}
-		pthread_mutex_unlock(&philo->info->mutex);
-		print_message(philo, TAKING_FORK, false);
 	}
 	else
 	{
-		pthread_mutex_lock(&philo->info->table.fork[philo->fork[L]]);
-		take_single_fork(philo, L);
-		pthread_mutex_lock(&philo->info->mutex);
-		if (quick_check(philo))
-		{
-			pthread_mutex_unlock(&philo->info->mutex);
+		if (safe_take_fork_odd(philo))
 			return ;
-		}
-		pthread_mutex_unlock(&philo->info->mutex);
-		print_message(philo, TAKING_FORK, false);
-		pthread_mutex_lock(&philo->info->table.fork[philo->fork[R]]);
-		take_single_fork(philo, R);
-		pthread_mutex_lock(&philo->info->mutex);
-		if (quick_check(philo))
-		{
-			pthread_mutex_unlock(&philo->info->mutex);
-			return ;
-		}
-		pthread_mutex_unlock(&philo->info->mutex);
-		print_message(philo, TAKING_FORK, false);
 	}
 }
